Adapt to the new Vivliostyle.js interface defined on `ui_separation` branch; extract entry code to a new module (`vivliostyle-viewer.js`) - See f8a9f5
