+++ src/html/vivliostyle-viewer.ejs
@@ -9,10 +9,9 @@
 <title>Vivliostyle.js viewer</title>
 <% if (development) { %>
     <script src="../node_modules/vivliostyle/src/vivliostyle.js"></script>
-    <script src="../node_modules/vivliostyle/src/vivliostyle/viewerapp.js"></script>
-    <script src="js/viewer-dev.js"></script>
+    <script src="js/main-dev.js"></script>
 <% } else { %>
-    <script src="js/viewer.js"></script>
+    <script src="js/main.js"></script>
 <% } %>
 <!--- install MathJax -->
 <script src="../src/MathJax/MathJax.js?config=MML_HTMLorMML"></script>
