+++ src/js/main-dev.js
@@ -1,0 +1,3 @@
+import vivliostyleViewer from "./vivliostyle-viewer";
+
+vivliostyleViewer.start(window.vivliostyle);
