+++ src/js/main.js
@@ -1,0 +1,4 @@
+import vivliostyle from "vivliostyle/build/vivliostyle.min.js";
+import vivliostyleViewer from "./vivliostyle-viewer";
+
+vivliostyleViewer.start(vivliostyle);
