--- src/js/viewer-dev.js
@@ -1,13 +1,0 @@
-function startViewer() {
-    var config = {
-        uaRoot: "resources/",
-        viewportElement: document.getElementById("vivliostyle-viewer-viewport")
-    };
-    vivliostyle.viewerapp.main(config);
-}
-
-if(window["__loaded"])
-    startViewer();
-else
-    window.onload = startViewer;
-
